pyproject.toml
[project]
name = "caudio"
version = "0.1.0"
description = "MCP server for audio generation (TTS) and transcription (STT)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.25.0",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/caudio"]
